requirements.txt.txt
flask
flask-cors
PyJWT
werkzeug
gunicorn